PR1863. Comments addressed. Types changed to allow scientific output

+++ src/ngraph/runtime/intelgpu/intelgpu_backend.hpp
@@ -76,9 +76,9 @@ private:
                                 const std::shared_ptr<Function> func,
                                 size_t time_compile,
                                 size_t time_call,
-                                size_t mem_before_call,
-                                size_t mem_after_compilation,
-                                size_t mem_after_call) const;
+                                double mem_before_call,
+                                double mem_after_compilation,
+                                double mem_after_call) const;
 
     bool m_profile_enable = false;
     long m_profile_lines_limit_count = 10;
